[FLINK-20625][pubsub,e2e] Add Javadocs, README and clean up
