--- flink-connectors/flink-connector-gcp-pubsub/src/main/java/org/apache/flink/streaming/connectors/gcp/pubsub/source/PubSubEnumeratorCheckpoint.java
@@ -1,7 +1,0 @@
-package org.apache.flink.streaming.connectors.gcp.pubsub.source;
-
-/** */
-public class PubSubEnumeratorCheckpoint {
-    //    could in theory hold a map of which projectSubscriptionName is assigned to what reader?!
-    //    see kafkasourceenumstate
-}
