--- flink-connectors/flink-connector-gcp-pubsub/src/main/java/org/apache/flink/streaming/connectors/gcp/pubsub/source/PubSubSplitState.java
@@ -1,4 +1,0 @@
-package org.apache.flink.streaming.connectors.gcp.pubsub.source;
-
-/** */
-public class PubSubSplitState extends PubSubSplit {}
