+++ flink-connectors/flink-connector-gcp-pubsub/src/main/java/org/apache/flink/streaming/connectors/gcp/pubsub/source/reader/PubSubSourceReader.java
@@ -1,4 +1,22 @@
-package org.apache.flink.streaming.connectors.gcp.pubsub.source;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.streaming.connectors.gcp.pubsub.source.reader;
 
 import org.apache.flink.api.connector.source.SourceReaderContext;
 import org.apache.flink.api.java.tuple.Tuple2;
@@ -7,6 +25,8 @@ import org.apache.flink.connector.base.source.reader.RecordEmitter;
 import org.apache.flink.connector.base.source.reader.RecordsWithSplitIds;
 import org.apache.flink.connector.base.source.reader.SingleThreadMultiplexSourceReaderBase;
 import org.apache.flink.connector.base.source.reader.synchronization.FutureCompletingBlockingQueue;
+import org.apache.flink.streaming.connectors.gcp.pubsub.source.split.PubSubSplit;
+import org.apache.flink.streaming.connectors.gcp.pubsub.source.split.PubSubSplitState;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,7 +36,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.Supplier;
 
-/** */
+/** The source reader to read from GCP Pub/Sub. */
 public class PubSubSourceReader<T>
         extends SingleThreadMultiplexSourceReaderBase<
                 Tuple2<T, Long>, T, PubSubSplit, PubSubSplitState> {
@@ -28,7 +48,6 @@ public class PubSubSourceReader<T>
             RecordEmitter<Tuple2<T, Long>, T, PubSubSplitState> recordEmitter,
             Configuration config,
             SourceReaderContext context) {
-        //		TODO: right super constructor?
         super(
                 elementsQueue,
                 new PubSubSourceFetcherManager<>(elementsQueue, splitReaderSupplier::get),
@@ -38,34 +57,42 @@ public class PubSubSourceReader<T>
     }
 
     @Override
-    protected void onSplitFinished(Map<String, PubSubSplitState> finishedSplitIds) {
-        //        try {
-        //            close();
-        //        } catch (Exception e) {
-        //            e.printStackTrace();
-        //        }
-    }
+    protected void onSplitFinished(Map<String, PubSubSplitState> finishedSplitIds) {}
 
     @Override
     public List<PubSubSplit> snapshotState(long checkpointId) {
-        // maybe match number of currently dealt-out splits (even though they are all the same)?
         return Arrays.asList(new PubSubSplit());
     }
 
+    /**
+     * Communicates with the {@link PubSubSourceFetcherManager} about the completion of a checkpoint
+     * so that messages received from GCP Pub/Sub can be acknowledged from a {@link
+     * PubSubSplitReader}.
+     *
+     * @param checkpointId the checkpoint ID.
+     * @throws Exception
+     */
     @Override
     public void notifyCheckpointComplete(long checkpointId) throws Exception {
-        LOG.info("Acknowledging Pub/Sub messages for checkpoint {}", checkpointId);
+        LOG.info("Acknowledging received GCP Pub/Sub messages for checkpoint {}", checkpointId);
         ((PubSubSourceFetcherManager<T>) splitFetcherManager).acknowledgeMessages();
     }
 
     @Override
     protected PubSubSplitState initializedState(PubSubSplit split) {
-        //		return new PubSubSplitState(split);
-        return null;
+        return new PubSubSplitState();
     }
 
+    /**
+     * Simply returns a new instance of {@link PubSubSplit} because GCP Pub/Sub offers no control of
+     * pulling messages beyond the configuration of project name and subscription name.
+     *
+     * @param splitId the split ID
+     * @param splitState the split state
+     * @return a fresh instance of {@link PubSubSplit}
+     */
     @Override
     protected PubSubSplit toSplitType(String splitId, PubSubSplitState splitState) {
-        return null;
+        return new PubSubSplit();
     }
 }
